requirements.txt
Django==1.10
mysqlclient==1.3.12
Pillow==5.0.0
djangorestframework==3.7.7
django-filter==1.1.0
